Move Reader to v0 since it includes v0-specific key-value encodings
